Start device units for uninitialised encrypted devices
This is necessary because the NixOS service that initialises the
filesystem depends on the appearance of the device unit.  Also, this
makes more sense to me: the device is ready; it's the filesystem
that's not, but taking care of that is the responsibility of the mount
unit.  (However, this ignores the fsck unit, so it's not perfect...)

Conflicts:
	rules/99-systemd.rules.in